requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

#sqlalchemy==2.0.23

pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2


yfinance>=0.2.65
akshare>=1.17.26


requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3


google-generativeai==0.3.2


python-dotenv==1.0.0
schedule==1.2.0


pytest==7.4.3
black==23.11.0
flake8==6.1.0
